refactor(datasetLinks): pull out Dataset navLink logic into seperate function
