+++ lib/components/dataset/DatasetHeader.js
@@ -7,7 +7,7 @@ import DatasetName from '../DatasetName'
 import Hash from '../Hash'
 import NavLinks from '../chrome/NavLinks'
 
-import { addActiveToLink } from '../../utils/links.js'
+import { addActiveToLink, datasetLinks } from '../../utils/links.js'
 import Base from '../Base'
 
 class DatasetHeader extends Base {
@@ -49,42 +49,11 @@ class DatasetHeader extends Base {
   // }
 
   template (css) {
-    const { url, datasetRef } = this.props
+    const { url, datasetRef, fromRegistry, isLocal } = this.props
     const { dataset } = datasetRef
     const pathname = this.props.location.pathname // this is why we need to wrap with 'withRouter'
 
-    const datasetLinksAll = [
-      {
-        name: 'Overview',
-        link: 'overview'
-      },
-      {
-        name: 'Viz',
-        link: 'viz'
-      },
-      {
-        name: 'Body',
-        link: 'body'
-      },
-      {
-        name: 'Meta',
-        link: 'meta'
-      },
-      {
-        name: 'Structure',
-        link: 'structure'
-      },
-      {
-        name: 'Transform',
-        link: 'transform'
-      },
-      {
-        name: 'History',
-        link: 'history'
-      }
-    ]
-    const keys = Object.keys(dataset)
-    var linkList = datasetLinksAll.filter(elem => elem.link === 'overview' || elem.link === 'body' || elem.link === 'history' || keys.includes(elem.link))
+    const linkList = datasetLinks(dataset, fromRegistry, isLocal)
 
     return (
       <div className={`border headerHeight ${css('wrap')}`}>
